JCF: a least-bad-option commit. Have two nearly identical GitHub Workflows for creating the slim externals images, one which is meant to be run scheduled, the other which is meant to be manually run and which takes inputs
In short, while I'd initially hoped to extend slim_externals.yaml by
letting users run it manually while also being able to override
variables (specifically the user account on mwts from which the
externals Spack packages get copied, and an optional suffix to add to
the produced image), this broke the scheduled version of the Workflow
since the scheduled version didn't have the variables set. I looked
into numerous workarounds, and while it _may_ be possible to do this
somehow, it's turning out to be laughably complicated. You'd think
that perhaps the scheduled version of the workflow would just take up
the default values of the input variables...but no. So instead, I'm
going to violate DRY and have two almost identical versions of the
same Workflow.